Loop url and is active tested and passed
