+++ public/scripts/models/loop.js
@@ -1,0 +1,14 @@
+(function(exports){
+
+  function Loop(url){
+    this.isActive = true;
+    this.url = url;
+
+    this.updateIsActive = function() {
+      this.isActive = !this.isActive;
+    }
+  }
+
+
+  exports.Loop = Loop;
+})(this);
