+++ views/partials/audio-recorder.ejs
@@ -17,21 +17,21 @@
       <button class='1'>Delete</button>
     </article>
 
-  </br>
+  <!-- </br> -->
 
   <div id="waveform2"></div>
     <article class='clip'>
       <button class='2'>Delete</button>
     </article>
 
-    </br>
+    <!-- </br> -->
 
       <div id="waveform3"></div>
     <article class='clip'>
       <button class='3'>Delete</button>
     </article>
 
-    </br>
+    <!-- </br> -->
 
     <div id="waveform4">
     </div>
@@ -39,28 +39,28 @@
       <button class='4'>Delete</button>
     </article>
 
-    </br>
+    <!-- </br> -->
 
     <div id="waveform5"></div>
     <article class='clip'>
       <button class='5'>Delete</button>
     </article>
 
-    </br>
+    <!-- </br> -->
 
     <div id="waveform6"></div>
     <article class='clip'>
       <button class='6'>Delete</button>
     </article>
 
-    </br>
+    <!-- </br> -->
 
       <div id="waveform7"></div>
     <article class='clip'>
       <button class='7'>Delete</button>
     </article>
 
-    </br>
+    <!-- </br> -->
 
     <div id="waveform8"></div>
     <article class='clip'>
