+++ views/partials/scripts.ejs
@@ -1,4 +1,7 @@
 <script src="//cdnjs.cloudflare.com/ajax/libs/wavesurfer.js/1.3.7/wavesurfer.min.js"></script>
+<script src='scripts/waveforms.js'></script>
 <script src='scripts/recorder.js'></script>
+<script src='scripts/models/loopFactory.js'></script>
+<script src='scripts/models/loop.js'></script>
 <script src='scripts/metronomeWorking.js'></script>
 <script src="http://cwilso.github.io/AudioContext-MonkeyPatch/AudioContextMonkeyPatch.js"></script>
